// dpbloom-client/src/app/core/api/model/createLectureDto.ts
/**
 * DPBloom | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateLectureDto { 
    title?: string;
    description?: string | null;
    content?: string;
    topicId?: string | null;
    contentLinks?: Array<string> | null;
    filePaths?: Array<string> | null;
}