// dpbloom-client/src/app/core/api/model/questionType.ts
/**
 * DPBloom | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const QuestionType = {
    SingleChoice: 'SingleChoice',
    MultipleChoice: 'MultipleChoice',
    OpenAnswer: 'OpenAnswer',
    TrueFalse: 'TrueFalse',
    Matches: 'Matches'
} as const;
export type QuestionType = typeof QuestionType[keyof typeof QuestionType];