// dpbloom-client/src/app/core/api/model/evaluationStrategy.ts
/**
 * DPBloom | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const EvaluationStrategy = {
    Best: 'Best',
    Last: 'Last'
} as const;
export type EvaluationStrategy = typeof EvaluationStrategy[keyof typeof EvaluationStrategy];